requirements.txt
django-ckeditor
django-celery-results
